+++ Magellan/Classes/Modules/Review/Reviews/ReviewTableViewCell.swift
@@ -1,0 +1,10 @@
+//
+/**
+* Copyright Soramitsu Co., Ltd. All Rights Reserved.
+* SPDX-License-Identifier: GPL-3.0
+*/
+
+
+import Foundation
+
+class ReviewTableViewCell: UITableViewCell {}
